refactor: 공통 레이아웃 컴포넌트(Header, Bottom) 분리 및 apply/FrontPage 헤더 경로 수정
