+++ src/components/Layout/Layout.module.css
@@ -1,0 +1,20 @@
+.layoutContainer {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+  width: 100%;
+  margin: 0 auto;
+  box-sizing: border-box;
+}
+
+.mainContent {
+  flex: 1;
+  width: 100%;
+  box-sizing: border-box;
+}
+
+.footer {
+  position: static;
+  width: 100%;
+  margin-top: auto;
+}
